PDJB-1169: Make landlord an abstract class
