+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/constants/enums/LandlordType.kt
@@ -1,0 +1,6 @@
+package uk.gov.communities.prsdb.webapp.constants.enums
+
+enum class LandlordType {
+    INDIVIDUAL,
+    ORGANISATION,
+}
